+++ CHANGES.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 - [BC-BREAK] Now `Request` trims and forces headers values to be string.
+- [BC-BREAK] Now `Request` throws `InvalidArgumentException` on non-string url.
 
 ## [0.3.0] - 2018-10-24